今日の作業終わり 画面が正常にクリアされない問題は解決せず。 描画処理を呼び出すごとにメモリの使用量が増える問題を解決している。ブレンドステートオブジェクトの生成を毎処理時呼び出さないようにした。

+++ EasyEngine_ver1/easy_engine.cpp
@@ -6,6 +6,8 @@
 #include "shader_manager.hpp"
 #include "texture_manager.hpp"
 
+#include "game.hpp"
+
 BEGIN_EGEG
 #ifdef _DEBUG
 int EasyEngine::create_cnt_;
@@ -34,6 +36,10 @@ void EasyEngine::run( const std::string& Platform )
     TextureManager::create( *this, p_texture_loader );
 
     // ゲーム実行
+    Game game;
+    if( game.initialize() == false ) return;
+    SystemManager::instance()->mainLoop( std::bind(&Game::update, &game, std::placeholders::_1) );
+    game.finalize();
 
     // マネージャーを破棄
     TextureManager::destroy( *this );
